+++ NEWS.md
@@ -6,6 +6,9 @@
 * New `httpbin_app()` end point: `/stream-bytes`, to send chunked
   random bytes.
 
+* New `httpbin_app()` end point: `/response-headers`, to set response
+  headers.
+
 * The web server now does not crash if the connection is closed by the
   client while delaying the response.